Scripts/requirements.txt
Pillow>=9.2.0
numpy>=1.23.0
numba>=0.56.0
pyaudio>=0.2.13